refactor exception handlers api-wide
